+++ src/lib/mongoBackend/mongoCreateSubscription.cpp
@@ -64,9 +64,9 @@ static void insertInCache
   const std::string&   tenant,
   const std::string&   servicePath,
   bool                 notificationDone,
-  long long            lastNotification,
-  long long            lastFailure,
-  long long            lastSuccess
+  double               lastNotification,
+  double               lastFailure,
+  double               lastSuccess
 )
 {
   //
@@ -209,7 +209,7 @@ std::string mongoCreateSubscription
 
   if (notificationDone)
   {
-    long long  lastNotification = (long long) getCurrentTime();
+    double lastNotification = getCurrentTime();
 
     setLastNotification(lastNotification, &b);
     setCount(1, &b);
